Fix EngagementsCarousel
