+++ assets/javascripts/kitten/karl/pages/homepage/index.js
@@ -9,7 +9,7 @@ import { Grid, GridCol } from '../../../components/grid/grid'
 import { ScreenConfig } from '../../../constants/screen-config'
 import COLORS from '../../../constants/colors-config'
 import TYPOGRAPHY from '../../../constants/typography-config'
-import EngagementsCarousel from './engagements-carousel'
+import { EngagementsCarousel } from './engagements-carousel'
 import { FeaturedProjects } from './featured-projects'
 import { PopularProjects } from './popular-projects'
 import { NewsBlock } from './news/components/news-block'
